cripple/backend/requirements.txt
Flask==2.2.3
PyYAML==6.0
